fixup! vk: review fixes

+++ ref/vk/shaders/brush.frag
@@ -23,7 +23,7 @@ layout(location=1) in vec3 vNormal;
 layout(location=2) in vec2 vTexture0;
 layout(location=3) in vec2 vLightmapUV;
 layout(location=4) in vec4 vColor;
-layout(location=5) flat in uint vLightingMode;
+layout(location=5) flat in float vLightingMode;
 
 layout(location=0) out vec4 outColor;
